Fixed build explosions
Stupid Unity feature... On build, Unity will only load assets that are
directly referenced by the project. The Detonator script was referencing
a material by calling Find("material-name"). Since this is not a direct
reference, Unity didn't import that material, so we got a null-pointer.
I fixed it using an equally stupid workaround: I referenced that
material and never used it.